I have added the files I have been working on
Start with parameters.m. Here you can alter your settings. Make sure that the output directory and the vsp3 filepaths are correct. To then run the VSPAERO, in vsp_analysis_body.py, on line 484 change it to the right .JSON file. Run the code, and it will give you the .csv file in the output directory you choose.